New event for January
This event will introduce attendees to the basics of blockchain and decentralized storage as provided on IPFS. It will be your opportunity ask questions and gain answers on the foundations of the Web3 Evolution. This event will be presented in both German and English. We will also have mini workshops that show you how to use Web3 to apply for roles on JobCrypt so if you're early you might get some free crypto. If not fear not we have more swag to give away through the course of the evening, so do join us.